Allow nonstandard ports when specifying network protocols.
There are usecases for running TFTP and HTTP on nonstandard ports.  This
patch allows you to specify nonstandard ports with the following syntax:

    set root=(http,$net_default_server,portno)
or
    set root=(http,,portno)

It also allows an initial : where a , should be for pxe: backwards compatibility

Rework net.c to comply with upstream requirements.

Add documentation for port protocol support.

This adds documentation for the new port section of the network
special drive support.